Massive refactor, many improvements
Legacy non-shader code paths removed from the OpenGL renderer
MFResource improvements
 - Custom resources
 - Better reference tracking
 - Many Fuji resources switched to use MFResource
Fuji resources _Destroy functions renamed to _Release
Created MFAllocator and MFPoolAllocator (Fuji STL allocators)
MFArray<> improvements
 - Lengths changed to size_t
 - Moving towards STL compliance, for( : ) loops.

+++ Fuji/Source/Asset/MFIntStringTable.cpp
@@ -232,7 +232,7 @@ MF_API MFIntStringTable *MFIntStringTable_CreateFromFileInMemory(const void *pMe
 			MFArray<MFString> row;
 			while(ReadCSVLine(pText, remaining, row))
 			{
-				for(int a=0; a<languages.size(); ++a)
+				for(size_t a=0; a<languages.size(); ++a)
 				{
 					if(!languages[a].IsEmpty())
 						columns[a].strings.push(row[a]);
@@ -250,7 +250,7 @@ MF_API MFIntStringTable *MFIntStringTable_CreateFromFileInMemory(const void *pMe
 
 	int numLanguages = 0;
 	int numStrings = 0;
-	for(int a=0; a<languages.size(); ++a)
+	for(size_t a=0; a<languages.size(); ++a)
 	{
 		if(languages[a].IsEmpty())
 			continue;
@@ -273,7 +273,7 @@ MF_API MFIntStringTable *MFIntStringTable_CreateFromFileInMemory(const void *pMe
 
 	MFStringCache *pStringCache = MFStringCache_Create(1024*1024);
 
-	for(int a=0, l=0; a<languages.size(); ++a)
+	for(size_t a=0, l=0; a<languages.size(); ++a)
 	{
 		if(languages[a].IsEmpty())
 			continue;
