+++ Fuji/Source/MFString.cpp
@@ -1471,7 +1471,7 @@ MFArray<MFString>& MFString::Split(MFArray<MFString> &output, const char *pDelim
 
 int MFString::Enumerate(const MFArray<MFString> keys, bool bCaseSensitive)
 {
-	for(int i=0; i<keys.size(); ++i)
+	for(size_t i=0; i<keys.size(); ++i)
 	{
 		if(bCaseSensitive ? Equals(keys[i]) : EqualsInsensitive(keys[i]))
 			return i;
